Add commands to Info and Goal panels (attempt 1)
*       s:cmddef() now takes in a flag aux, telling it whether the
        command is being defined for the main panel or an auxiliary
        panel; we thread that argument through s:commands().

*       When define these commands in coqtail#register(), we only define
        them for the main buffer, with the aux flag set to 0, behaving
        as it used to. We cannot yet register the commands for the other
        panels, since they haven't been created (they're only created
        upon :CoqStart).

*       In coqtail#start() (called by :CoqStart), once the info and goal
        buffers are created, we define the commands with the aux flag
        set to 1. This will wrap each command definition to to
        temporarily switch to the main panel, where the command is
        actually defined, execute the command, and switch back.

Note that we only need to do this because a lot of commands seem to rely
on variables local to the main buffer. This seems to work for now, but
might limit what kinds of commands we can define, or at least make the
logic of some commands really confusing (since the commands themselves
will also be switching between buffers to do what they need to do).

An alternative approach would be to refactor all buffer-local variables
to some "environment" dictionary shared between the main, info, and goal
buffers, but that would require bigger changes.