Remove buffer switch from s:cmddef()
*       Remove aux argument from s:cmddef() and define_commands()
*       Pass main buffer number instead of bufnr('%') in s:call()
*       Call setbufvar(b:coqtail_panel_bufs.main, &modifiable, 0)
        instead of using setlocal in s:call()

Note that in coqtail#defaultCB(), we still use a:buf instead of
hardcoding b:coqtail_panel_bufs.main, since it's possible that we've
switched to a non-coqtail buffer where b:coqtail_panel_bufs isn't
defined. a:buf should be the main buffer number we passed to Python
in s:call().