Clarify changedtick and buffer
changedtick and buffer can be out of sync if the user concurrently edits
the buffer while Coq is processing. Currently, this is prevented by
locking the buffer, so it's not really an issue. However, for future
development (e.g. full async processing), it's desirable to clearly
differentiate the synchronized pair of changedtick and buffer from the
functions that read those values from Vim.